requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
Django==3.2.11
django-jsonfield==1.4.1
djangorestframework==3.13.1
drf-yasg==1.20.0
html-to-json==2.0.0
pyOpenSSL==21.0.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
webdriver-manager==3.5.2